Move back to own package
ASF contribution was "rejected" years ago mainly because of the Bouncy
Castle dependency. The package change was motivated only by this at this
time. This project should not have been released as such. IOW. 1.1 is
legally broken and and will be "removed" from where possible as soon as
the next version is out.